Move diagnostics to a more logical place. Before the diagnostics for T=10 were printed for T<=8 but named grid_10. This causes problems at restart.
Added automatic stop to prevent going over runtime. This will prevent corrupted files appearing. Diagnostics are kept in the same place regardless of the starting point (e.g. at T=10,20,30,40, and with restart at T=24 diagnostics at T=30,40, not T=34,44). This allows are large print step to be used as the final state is saved anyway for the restart.